popout.php: Move sub function and adjust expected src url
This moves the subscribe button function to the page itself, so that the entire rachni.js file does not need to be loaded. Eventually, this can be changed back once rachni.js has been sufficiently cleaned up to not generate errors in a vacuum. This also changes the expected src url from using a / delimiter to & to indicate the stream title.