Fix administrative functions
This update migrates the administrative check to a more secure function,
using the user table in the database to verify if a user is an admin.
There is still no way of adding an administrator manually, but this
will be addresed in future updates.

Additionally, the admin page in settings has been updated to allow for
in-line editing of the user table, if the need arises. There are meny
fields currently missing from the table, and the display is busted,
but it's a good start.