Large update to allow custom offline stream images
This update allows users to submit their own image to be used if a
channel page is browsed to while the user is offline. It can be uploaded
on the user settings page. .jpg, .png, and .gif are all allowed. As a
side effect of this update, and due to the poor error handling of
video.js, any video.js errors are suppressed in favor of showing the
offline image.