Refresh player when stream goes from offline -> online
This update will detect if a stream is either offline or online, and
will automatically refresh the player to start the stream playing if
the streamer goes live while the page was open. I will probably add
this as an option in user settings to enable/disable, but the proof of
concept is complete and working correctly. This also increases the
heartbeat timeout a bit, as it was a little noisy before.