Fix scroll position after fullscreen event
This small fix corrects a bug where the scrollbar position would be
moved after the video.js player was set to fullscreen. A small timeout
is required because the video.js event is not triggered after the
fullscreen change is complete, but rather during the process itself. I
chose to set the scrollbar to the bottom as most users would want to
see the most recent messages after the chat had been hidden.