Update resetPlayer function
This updates allows the resetPlayer function to handle all the functions
that need to happen when a stream is set back live, including removing
the poster and setting the live_status variable to true. This has no
functional difference, but allows the Reset Player button at the top of
the player pages to function properly for an offline channel (though
manually needed to reset the player should not be required anymore).